pyproject.toml
[project]
name = "nnja-ai-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp==2.12.4",
    "fuzzywuzzy>=0.18.0",
    "google-genai>=1.43.0",
    "matplotlib>=3.10.7",
    "nnja-ai>=1.0.0",
    "pyqt6>=6.9.1",
]
